refactor: move related articles to article
moving related articles to artilce paves the way to make the slice package only related to
templating
