+++ packages/article/__tests__/android/article.android.test.js
@@ -1,5 +1,7 @@
 import shared from "../shared";
+import relatedArticlesShared from "../related-articles-shared";
 
 describe("Article test on android", () => {
   shared();
+  relatedArticlesShared();
 });
