+++ packages/article/__tests__/ios/article.ios.test.js
@@ -1,5 +1,7 @@
 import shared from "../shared";
+import relatedArticlesShared from "../related-articles-shared";
 
 describe("Article test on ios", () => {
   shared();
+  relatedArticlesShared();
 });
