+++ packages/article/package.json
@@ -38,6 +38,8 @@
   },
   "homepage": "https://github.com/newsuk/times-components#readme",
   "devDependencies": {
+    "@storybook/addon-actions": "3.3.11",
+    "@storybook/addons": "3.3.11",
     "@storybook/react-native": "3.3.11",
     "@times-components/eslint-config-thetimes": "0.1.3",
     "@times-components/jest-configurator": "0.0.22",
@@ -49,6 +51,7 @@
     "enzyme-to-json": "3.3.0",
     "eslint": "4.9.0",
     "jest": "21.2.1",
+    "jest-context": "2.1.0",
     "jest-styled-components": "4.9.0",
     "prettier": "1.8.2",
     "react": "16.2.0",
@@ -72,6 +75,7 @@
     "@times-components/provider": "0.22.2",
     "@times-components/pull-quote": "0.2.7",
     "@times-components/responsive-styles": "0.2.15",
+    "@times-components/slice": "0.0.4",
     "@times-components/tracking": "0.6.10",
     "lodash.get": "4.4.2",
     "prop-types": "15.6.0"
