+++ packages/article/related-articles/fixtures/single-related-article-no-image.json
