+++ packages/article/related-articles/heading.js
@@ -1,11 +1,7 @@
-import { StyleSheet } from "react-native";
+import React from "react";
+import { StyleSheet, View, Text } from "react-native";
 
-const sharedStyle = {
-  container: {
-    marginTop: 10,
-    paddingLeft: 10,
-    paddingRight: 10
-  },
+const styles = StyleSheet.create({
   titleContainer: {
     borderStyle: "solid",
     borderBottomColor: "#dbdbdb",
@@ -21,11 +17,13 @@ const sharedStyle = {
     fontFamily: "TimesModern-Bold",
     fontSize: 26,
     color: "#333"
-  },
-  cardContainer: {
-    paddingBottom: 10,
-    paddingTop: 10
   }
-};
+});
 
-export default sharedStyle;
+const RelatedArticlesHeading = () => (
+  <View style={styles.titleContainer}>
+    <Text style={styles.title}>Related links</Text>
+  </View>
+);
+
+export default RelatedArticlesHeading;
