+++ packages/slice/__tests__/android/slice.test.js
@@ -1,4 +1,4 @@
-import shared from "../shared";
+import shared from "@times-components/article/__tests__/related-articles-shared.js";
 
 describe("Slice test on android", () => {
   shared();
